Added reader to the system

+++ reader.h
@@ -1,2 +1,4 @@
+#include "avr/io.h"
+
 void init_ADC();
-uint16_t read_ADC(uint16_t _channel);
+uint16_t read_ADC(uint8_t _channel);
